+++ NEWS.md
@@ -7,6 +7,10 @@
 -   Implemented a new feature to add density line contours using `plot_density_contour`.
 -   Implemented the conditional use of `raster.dpi` to Seurat versions higher or equal to 4.1.0.
 
+## `SCpubr::do_EnrichmentHeatmap()`
+-   Implemented a bug fix for internal checks in the function.
+-   Added `min.cutoff` and `max.cutoff` parameter to effectively subset the color scale and remove the effect of extreme outliers in the matrix.
+
 ## `SCpubr::do_FeaturePlot()`
 -   Implemented a new feature to add density line contours using `plot_density_contour`.
 -   Implemented the conditional use of `raster.dpi` to Seurat versions higher or equal to 4.1.0.